+++ garnish/__init__.py
@@ -1,3 +1,5 @@
+# Copyright 2019-2019 the garnish authors. See copying.md for legal info.
+
 from .garnish import extend_garnish
 from .utils import extension
